+++ src/main/java/poopprojekat/studentska_sluzba/UserController.java
@@ -5,30 +5,35 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 //INCLUDES:
-// /get_user_role -returns role of user with given username todo and password da se doda da moze samo preko username-a da se pretrazuje
-// /add_user?username=&password=&year= -adds new user with given username, password todo and year nije svaki user student pa da ima indeks i godinu upisa
-// /update_user?username=&new_username=&new_password=&new_role= -updates user with given username, sets new given values
-// /delete_user?unique_id= -deletes ueser wiht given id
+// /get_user?token=&username=&password= -returns user with given username todo and password da se doda da moze samo preko username-a da se pretrazuje
+// /add_user?token=&username=&password=&year= -adds new user with given username, password todo and year nije svaki user student pa da ima indeks i godinu upisa
+// /update_user?token=&username=&new_username=&new_password=&new_role= -updates user with given username, sets new given values
+// /delete_user?token=&unique_id= -deletes user with given id
 
 @RestController
 public class UserController {
 
-    @GetMapping("/get_user_role")
-    public String get_user_role(@RequestParam("username") String username,
-                                @RequestParam("password") String password){
+    @GetMapping("/get_user")
+    public String[] get_user(@RequestParam("token") long token,
+                             @RequestParam("username") String username,
+                             @RequestParam("password") String password){
 
-        /*String tmp = Database.GetUser(username, password);
+        if (!(Log_in_Controller.contains_user(token)[0]).equals("Admin")) return null;
 
-        if (tmp == null) return "User don't exist";
-        return tmp;*/
-        return null;
+        String[] tmp = Database.GetUser(username, password);
+
+        if (tmp == null) return new String[]{"User don't exist", null};
+        return tmp;
     }
 
     @GetMapping("/add_user")
-    public String add_user(@RequestParam("username") String username,
+    public String add_user(@RequestParam("token") long token,
+                           @RequestParam("username") String username,
                            @RequestParam("password") String password,
                            @RequestParam("role") String role,
                            @RequestParam("year") int year){
+
+        if (!(Log_in_Controller.contains_user(token)[0]).equals("Admin")) return null;
 
         User user = new User(username, password, role);
 
@@ -43,10 +48,13 @@ public class UserController {
     }
 
     @GetMapping("/update_user")
-    public String update_user(@RequestParam("username") String username,
+    public String update_user(@RequestParam("token") long token,
+                              @RequestParam("username") String username,
                               @RequestParam("new_username") String new_username,
                               @RequestParam("new_password") String new_password,
                               @RequestParam("new_role") String new_role){
+
+        if (!(Log_in_Controller.contains_user(token)[0]).equals("Admin")) return null;
 
         User updated_user = new User(new_username, new_password, new_role);
         try{
@@ -59,7 +67,11 @@ public class UserController {
     }
 
     @GetMapping("/delete_user")
-    public String delete_user(@RequestParam("unique_id") String id){
+    public String delete_user(@RequestParam("token") long token,
+                              @RequestParam("unique_id") String id){
+
+        if (!(Log_in_Controller.contains_user(token)[0]).equals("Admin")) return null;
+
         try{
             Database.DeleteUser(id);
             return "User successfully deleted";
